+++ editors/jetbrains/src/main/kotlin/io/agnix/jetbrains/lsp/AgnixLspServerSupportProvider.kt
@@ -3,6 +3,7 @@ package io.agnix.jetbrains.lsp
 import com.intellij.openapi.project.Project
 import com.redhat.devtools.lsp4ij.LanguageServerFactory
 import com.redhat.devtools.lsp4ij.client.LanguageClientImpl
+import com.redhat.devtools.lsp4ij.client.features.LSPClientFeatures
 import com.redhat.devtools.lsp4ij.installation.ServerInstaller
 import com.redhat.devtools.lsp4ij.server.StreamConnectionProvider
 
@@ -24,6 +25,10 @@ class AgnixLspServerSupportProvider : LanguageServerFactory {
 
     override fun createServerInstaller(): ServerInstaller {
         return AgnixLspServerInstaller()
+    }
+
+    override fun createClientFeatures(): LSPClientFeatures {
+        return AgnixLspClientFeatures()
     }
 
     /**